requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-form==3.4
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.2
grequests==0.6.0
gunicorn==20.1.0
html5lib==1.1
idna==3.2
iniconfig==1.1.1
monkey==0.1
multidict==6.0.2
mysqlclient==2.1.0
nose==1.3.7
nti.nose-traceback-info==1.0.2
packaging==21.3
Pillow==8.3.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pyshorteners==1.0.1
pytest==7.1.1
python-dotenv==0.19.2
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tomli==2.0.1
urllib3==1.26.6
webencodings==0.5.1
whitenoise==6.0.0
yarl==1.8.1
zope.event==4.5.0
zope.exceptions==4.5
zope.interface==5.4.0
